Require callers of `Gagat.configure(…)` to specify the styleable object
This reverts to how it worked prior to f577b50.

I realized that it’s not very nice to force people to bloat their
(probably massive) view controllers even more by using the window’s root
view controller as the styleable object. It’s still possible to use the
root view controller as the styleable object - that’s what we do
in the example app - but it’s no longer a requirement.

Structuring the API in this way also makes it a bit clearer that Gagat
calls `toggleActiveStyle()` solely on the styleable object, not on the
entire view controller hierarchy.