Remove redundant style attributes from Example app
In 178e79b we introduced the `StyleableNavigationController` which
manages the appearance of the navigation bar and the status bar. That
means those attributes were no longer needed in
`ArchiveTableViewController` and thus they were removed in this commit.
Since the only thing left in the `Style` struct was `TableViewStyle`, I
decided to remove `Style` and rename `TableViewStyle` to `Style`.