Replace pre-built binaries instructions with subproject instructions
Since Gagat is built using Swift, using pre-built binaries is not the
best recommendation due to the lack of ABI stability between different
Swift versions. Adding Gagat as a subproject is a safer option since the
framework(s) are then built from source using the same Xcode version as
the app that is linking against them.